+++ jetzt/update_metadata.py
@@ -9,6 +9,7 @@ metadata = {}
 dep_type = 'PROD'
 package = ''
 package_with_version = ''
+action = ''
 
 ''' Read existing metadata if available. '''
 if os.path.exists(metadata_filename):
@@ -25,12 +26,27 @@ for arg in sys.argv:
             package = value
         elif key == 'package_with_version':
             package_with_version = value
+        elif key == 'action':
+            action = value
         else:
             metadata[key] = value
     argv_count += 1
 
-''' Installing a new dependency. '''
-if package != '':
+dependencies = 'dependencies'
+if dep_type == 'DEV':
+    dependencies = 'dev_dependencies'
+
+if action == 'REMOVE':
+    ''' Removing a dependency '''
+    if dependencies in metadata:
+        if isinstance(metadata[dependencies], dict):
+            try:
+                del metadata[dependencies][package]
+            except KeyError as e:
+                print(f"Could not remove dependency {package} from jetzt_metadata.json ({dependencies}/{package}), please remove manually.")
+
+elif action == 'INSTALL':
+    ''' Installing a new dependency. '''
     installed_package, installed_version = re.split("==|>=", package_with_version)  # package==0.13.2
     # Pinned version
     if '==' in package:
@@ -38,20 +54,12 @@ if package != '':
     else:
         installed_version = f">={installed_version}"
 
-    if dep_type == 'PROD':
-        if 'dependencies' in metadata:
-            if isinstance(metadata['dependencies'], dict):
-                metadata['dependencies'][installed_package] = installed_version
-        else:
-            metadata['dependencies'] = {}
-            metadata['dependencies'][installed_package] = installed_version
+    if dependencies in metadata:
+        if isinstance(metadata[dependencies], dict):
+            metadata[dependencies][installed_package] = installed_version
     else:
-        if 'dev_dependencies' in metadata:
-            if isinstance(metadata['dev_dependencies'], dict):
-                metadata['dev_dependencies'][installed_package] = installed_version
-        else:
-            metadata['dev_dependencies'] = {}
-            metadata['dev_dependencies'][installed_package] = installed_version
+        metadata[dependencies] = {}
+        metadata[dependencies][installed_package] = installed_version
 
 ''' Dump updated metadata. '''
 with open(metadata_filename, 'w') as dump: