+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2019-04-14
+
+### Added
+
+- Added ability to remove dependencies, with `jetzt --remove`.
+- Added ability to generate standard `requirements.txt` and `requirements-dev.txt` based on `jetzt_metadata.json`, with `jetzt --create-requirements`.
+
+### Changed
+
+### Removed
+
+### Fixed
+
+
 ## [0.4.1] - 2019-04-13
 
 ### Added